[TR-274] refactor: add response mappers

+++ data/src/main/java/ru/radiationx/data/entity/response/auth/OtpInfoResponse.kt
@@ -2,12 +2,11 @@ package ru.radiationx.data.entity.response.auth
 
 import com.squareup.moshi.Json
 import com.squareup.moshi.JsonClass
-import java.util.*
 
 @JsonClass(generateAdapter = true)
 data class OtpInfoResponse(
     @Json(name = "code") val code: String,
     @Json(name = "description") val description: String,
-    @Json(name = "expiredAt") val expiredAt: Long,
-    @Json(name = "remainingTime") val remainingTime: Long
+    @Json(name = "expiredAt") val expiredAt: Int,
+    @Json(name = "remainingTime") val remainingTime: Int
 )
